+++ jade/main.jade
@@ -6,6 +6,29 @@ html
     meta(http-equiv="X-UA-Compatible", content="IE=edge,chrome=1")
 
     link(rel='stylesheet', href='css/main.css')
+    script( rc='components/stacktrace/stacktrace.js')
+
+    script.
+        //- Catch all errors
+        process.on('uncaughtException', function(err) {
+            console.log(err);
+            console.log(err.fileName);
+        });
+
+        //- Prevent namespace conflict with node require
+        window.requireNodeBackup = window.require;
+        window.require = undefined;
+        window.requireNode = function(moduleName){
+            var requirejs = require;
+            window.require = window.requireNodeBackup;
+            var module = require(moduleName);
+            window.require = requirejs
+            return module;
+        };
+
+
+
+    script(data-main="js/config", src='components/requirejs/require.js')
 
 body(ontouchstart).clear.md-mode
     .container
@@ -155,17 +178,17 @@ body(ontouchstart).clear.md-mode
             input.newKey.pull-left(placeholder="key")
             input.newValue.pull-right(placeholder="value")
 
-    script(src='components/jquery/jquery.js')
-    script(src='components/underscore/underscore.js')
-    script(src='components/backbone-deep-model/lib/underscore.mixin.deepExtend.js')
-    script(src='components/backbone/backbone.js')
-    script(src='components/backbone-deep-model/src/deep-model.js')
-    script(src='components/moment/moment.js')
-    script(src='components/nprogress/nprogress.js')
-    script(src='components/floatLabel/jquery.FloatLabel.js')
-    script(src='components/ace-builds/src-noconflict/ace.js')
-    script(src='components/keymaster/keymaster.js')
-    script(src='js/namespaces.js')
-    script(src='js/helpers.js')
-    script(src='js/main.js')
-    script(src='http://localhost.com:35729/livereload.js')
+    //- script(src='components/jquery/jquery.js')
+    //- script(src='components/underscore/underscore.js')
+    //- script(src='components/backbone-deep-model/lib/underscore.mixin.deepExtend.js')
+    //- script(src='components/backbone/backbone.js')
+    //- script(src='components/backbone-deep-model/src/deep-model.js')
+    //- script(src='components/moment/moment.js')
+    //- script(src='components/nprogress/nprogress.js')
+    //- script(src='components/floatLabel/jquery.FloatLabel.js')
+    //- script(src='components/ace-builds/src-noconflict/ace.js')
+    //- script(src='components/keymaster/keymaster.js')
+    //- script(src='js/namespaces.js')
+    //- script(src='js/helpers.js')
+    //- script(src='js/main.js')
+
