+++ js/config.js
@@ -1,0 +1,20 @@
+require.config({
+    baseUrl: "./js",
+    deps: ["main"],
+    paths: {
+        "jquery"          : "../components/jquery/jquery",
+        "underscore"      : "../components/underscore-amd/underscore",
+        "underscore-deep" : "../components/backbone-deep-model/lib/underscore.mixin.deepExtend",
+        "backbone"        : "../components/backbone-amd/backbone",
+        "backbone-deep"   : "../components/backbone-deep-model/src/deep-model",
+        "floatLabel"      : "../components/floatLabel/jquery.floatLabel",
+        "nprogress"       : "../components/nprogress/nprogress",
+        "moment"          : "../components/moment/moment",
+        "ace"             : "../components/ace-builds/src-noconflict/ace",
+        "keymaster"       : "../components/keymaster/keymaster",
+        "app"             : "app",
+        "db"              : "db",
+        "settings"        : "settings",
+        "events"          : "events",
+    },
+});
