+++ js/db.js
@@ -1,0 +1,8 @@
+define([], function(){
+    return {
+        module: requireNode('mongojs'),
+        client: null,
+        collection: null,
+    }
+
+});
