+++ js/gui.js
@@ -1,0 +1,3 @@
+define([], function(){
+    return requireNode('nw.gui');
+});
