+++ js/module_template.js
@@ -1,0 +1,12 @@
+define([
+    'jquery',
+    'underscore',
+    'backbone',
+    'helpers'
+], function($, _, BB, Helpers){
+
+
+    return {
+
+    }
+});
