+++ js/settings.js
@@ -1,0 +1,13 @@
+define([], function(){
+    return {
+        Root : '/',
+        Auth : {
+            host       : 'localhost.com',
+            port       : 27017,
+            username   : '',
+            psw        : '',
+            database   : '',
+            collection : ''
+        }
+    }
+});
