+++ main.html
@@ -5,6 +5,29 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
     <link rel="stylesheet" href="css/main.css">
+    <script rc="components/stacktrace/stacktrace.js"></script>
+    <script>
+      //- Catch all errors
+      process.on('uncaughtException', function(err) {
+          console.log(err);
+          console.log(err.fileName);
+      });
+      
+      //- Prevent namespace conflict with node require
+      window.requireNodeBackup = window.require;
+      window.require = undefined;
+      window.requireNode = function(moduleName){
+          var requirejs = require;
+          window.require = window.requireNodeBackup;
+          var module = require(moduleName);
+          window.require = requirejs
+          return module;
+      };
+      
+      
+      
+    </script>
+    <script data-main="js/config" src="components/requirejs/require.js"></script>
   </head>
 </html>
 <body ontouchstart class="clear md-mode">
@@ -135,18 +158,4 @@
       <input placeholder="value" class="newValue pull-right">
     </div>
   </script>
-  <script src="components/jquery/jquery.js"></script>
-  <script src="components/underscore/underscore.js"></script>
-  <script src="components/backbone-deep-model/lib/underscore.mixin.deepExtend.js"></script>
-  <script src="components/backbone/backbone.js"></script>
-  <script src="components/backbone-deep-model/src/deep-model.js"></script>
-  <script src="components/moment/moment.js"></script>
-  <script src="components/nprogress/nprogress.js"></script>
-  <script src="components/floatLabel/jquery.FloatLabel.js"></script>
-  <script src="components/ace-builds/src-noconflict/ace.js"></script>
-  <script src="components/keymaster/keymaster.js"></script>
-  <script src="js/namespaces.js"></script>
-  <script src="js/helpers.js"></script>
-  <script src="js/main.js"></script>
-  <script src="http://localhost.com:35729/livereload.js"></script>
 </body>
