+++ package.json
@@ -6,10 +6,17 @@
   "private": true,
   "node-remote": "<local>",
   "no-edit-menu": true,
+  "author": "Jose Browne <josebrowne@gmail.com>",
+  "manifestUrl": "http://localhost:3000/package.json",
+  "packages": {
+    "mac": "http://localhost:3000/builds/releases/updapp/mac/updapp.dmg",
+    "win": "http://localhost:3000/builds/releases/updapp/win/updapp.zip",
+    "linux32": "http://localhost:3000/builds/releases/updapp/linux32/updapp.tar.gz"
+  },
   "window": {
     "title": "MongoCMS",
     "position": "center",
-    "toolbar": false,
+    "toolbar": true,
     "frame": true,
     "width": 1270,
     "height": 800,
@@ -19,19 +26,22 @@
   },
   "dependencies": {
     "mongodb": "~1",
-    "mongojs": "~0.13.0"
+    "mongojs": "~0.13.1"
   },
   "devDependencies": {
     "grunt": "~0.4.5",
     "grunt-develop": "~0.4.0",
-    "request": "~2.36.0",
+    "request": "~2.39.0",
     "grunt-contrib-watch": "~0.6.1",
     "grunt-contrib-sass": "~0.7.3",
-    "grunt-contrib-concat": "~0.4.0",
-    "grunt-contrib-uglify": "~0.5.0",
-    "grunt-autoprefixer": "~0.8.1",
+    "grunt-contrib-concat": "~0.5.0",
+    "grunt-contrib-uglify": "~0.5.1",
+    "grunt-autoprefixer": "~0.8.2",
     "grunt-contrib-jade": "~0.12.0",
     "grunt-contrib-compass": "~0.9.0",
-    "grunt-node-webkit-builder": "~0.1.21"
+    "grunt-node-webkit-builder": "~0.1.21",
+    "grunt-contrib-compress": "~0.10.0",
+    "grunt-contrib-copy": "~0.5.0",
+    "grunt-contrib-clean": "~0.6.0"
   }
 }
